Rohitmagdum (#11)
* Added the possible scenes enums.

Tried to break the scene into as much granularity as possible. This is
not enough but is a start. Further more values will be added.

* converted tabs into spaces and removed trailing numbers after enums in the last commit

* Just adding my code and framework I use.

Guys just adding the framework I have been using along with one test
project. If anything from this is helpful please use. Any comments
related to the same are most welcome.

* Added handling of vertex, normals, tex coord buffers

Changed the handling of Vertex, Normals and Tex Coords buffer. Keeping
them member of derived class and base class agnostic of the data.